requirements.txt
polygon-api-client>=1.13
anthropic>=0.49
schedule>=1.2
python-dotenv>=1.0
gitpython>=3.1
